+++ frame/staking/src/lib.rs
@@ -1204,7 +1204,7 @@ pub mod pallet {
 
 	/// The last planned session scheduled by the session pallet.
 	///
-	/// This is basically in sync with the call to [`SessionManager::new_session`].
+	/// This is basically in sync with the call to [`pallet_session::SessionManager::new_session`].
 	#[pallet::storage]
 	#[pallet::getter(fn current_planned_session)]
 	pub type CurrentPlannedSession<T> = StorageValue<_, SessionIndex, ValueQuery>;
@@ -1511,8 +1511,8 @@ pub mod pallet {
 		/// The dispatch origin for this call must be _Signed_ by the stash, not the controller.
 		///
 		/// Use this if there are additional funds in your stash account that you wish to bond.
-		/// Unlike [`bond`] or [`unbond`] this function does not impose any limitation on the amount
-		/// that can be added.
+		/// Unlike [`bond`](Self::bond) or [`unbond`](Self::unbond) this function does not impose any limitation
+		/// on the amount that can be added.
 		///
 		/// Emits `Bonded`.
 		///
@@ -1849,7 +1849,7 @@ pub mod pallet {
 		/// The dispatch origin must be Root.
 		///
 		/// # <weight>
-		/// Same as [`set_validator_count`].
+		/// Same as [`Self::set_validator_count`].
 		/// # </weight>
 		#[pallet::weight(T::WeightInfo::set_validator_count())]
 		pub fn increase_validator_count(
@@ -1866,7 +1866,7 @@ pub mod pallet {
 		/// The dispatch origin must be Root.
 		///
 		/// # <weight>
-		/// Same as [`set_validator_count`].
+		/// Same as [`Self::set_validator_count`].
 		/// # </weight>
 		#[pallet::weight(T::WeightInfo::set_validator_count())]
 		pub fn scale_validator_count(origin: OriginFor<T>, factor: Percent) -> DispatchResult {
@@ -3047,6 +3047,57 @@ impl<T: Config> frame_election_provider_support::ElectionDataProvider<T::Account
 		now.saturating_add(
 			until_this_session_end.saturating_add(sessions_left.saturating_mul(session_length)),
 		)
+	}
+
+	#[cfg(any(feature = "runtime-benchmarks", test))]
+	fn add_voter(voter: T::AccountId, weight: VoteWeight, targets: Vec<T::AccountId>) {
+		use sp_std::convert::TryFrom;
+		let stake = <BalanceOf<T>>::try_from(weight).unwrap_or_else(|_| {
+			panic!("cannot convert a VoteWeight into BalanceOf, benchmark needs reconfiguring.")
+		});
+		<Bonded<T>>::insert(voter.clone(), voter.clone());
+		<Ledger<T>>::insert(
+			voter.clone(),
+			StakingLedger {
+				stash: voter.clone(),
+				active: stake,
+				total: stake,
+				unlocking: vec![],
+				claimed_rewards: vec![],
+			},
+		);
+		Self::do_add_nominator(
+			&voter,
+			Nominations { targets: targets, submitted_in: 0, suppressed: false },
+		);
+	}
+
+	#[cfg(any(feature = "runtime-benchmarks", test))]
+	fn add_target(target: T::AccountId) {
+		let stake = MinValidatorBond::<T>::get() * 100u32.into();
+		<Bonded<T>>::insert(target.clone(), target.clone());
+		<Ledger<T>>::insert(
+			target.clone(),
+			StakingLedger {
+				stash: target.clone(),
+				active: stake,
+				total: stake,
+				unlocking: vec![],
+				claimed_rewards: vec![],
+			},
+		);
+		Self::do_add_validator(
+			&target,
+			ValidatorPrefs { commission: Perbill::zero(), blocked: false },
+		);
+	}
+
+	#[cfg(any(feature = "runtime-benchmarks", test))]
+	fn clear() {
+		<Bonded<T>>::remove_all(None);
+		<Ledger<T>>::remove_all(None);
+		<Validators<T>>::remove_all(None);
+		<Nominators<T>>::remove_all(None);
 	}
 
 	#[cfg(any(feature = "runtime-benchmarks", test))]